djangotutorial/requirements.txt
Django==4.2.9
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.7
setuptools==78.1.0
requests==2.32.3
gunicorn==23.0.0
psycopg[binary]==3.2.9